Fixed code->char and char->code
They now detect invalid input and inform the caller.

Valid input is a code (or corresponding char) above 32 and not
equal to 127 (the delete key).

NOTE: Should probably add an upper bound, too. Should this be
implementation-dependant (ie, detect the char encoding of the
current Lisp), or should we assume that UTF-8 is supported?